# openapi.yaml
...
openapi: 3.0.3
info:
  title: "GitHub Insight Engine API"
  description: |
    REST API for GitHub data collection, analysis, and insights generation.
    
    **Authentication**: JWT tokens obtained via `/auth/login`
    **Data Flow**: Collect → Store → Analyze → Retrieve
    **Rate Limits**: Standard GitHub API rate limits apply during collection
    **Automatic Scheduling**: Built-in scheduling service with automatic token management
  version: 0.1.90
  contact:
    name: Github
    url: https://github.com/octocat/gh-insight-engine
  license:
    name: MIT
    url: https://github.com/octocat/gh-insight-engine/blob/main/LICENSE
servers:
  - url: https://github.lucas-blanchard.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
  - {}
paths:
...
  /repositories/{username}:
    post:
      summary: Collect repository data
      description: |
        Start repository collection process. Returns immediately with status 202 Accepted.
        Use GET /repositories/{username}/status to monitor progress.
      operationId: collectRepositoriesData
      tags:
        - Repositories
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '202':
          description: Repository collection started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryCollectionStartedResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '403':
          description: Can only collect own repository data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: GitHub API rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "RATE_LIMIT_ERROR"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "GitHub API rate limit exceeded. Please try again later or contact GitHub Support if the issue persists."
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"

    get:
      summary: Get user repositories
      description: Retrieve stored repositories and analytics
      operationId: getUserRepositories
      tags:
        - Repositories
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '200':
          description: Repositories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoriesResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '404':
          description: No repositories found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete user repositories
      description: Delete all stored repositories for a user from the database
      operationId: deleteUserRepositories
      tags:
        - Repositories
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '200':
          description: Repositories deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryDeletionResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '403':
          description: Can only delete own repository data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No repositories found to delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: GitHub API rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "RATE_LIMIT_ERROR"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "GitHub API rate limit exceeded. Please try again later or contact GitHub Support if the issue persists."
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
  /repositories/{username}/status:
    get:
      summary: Get repository collection status
      description: Get the status of repository collection for a user
      operationId: getCollectionStatus
      tags:
        - Repositories
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernameParam'
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryStatusResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error type
                    example: "Unauthorized"
                  message:
                    type: string
                    description: Human-readable error message
                    example: "Please log in to access this resource"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-07-28T07:36:05.494Z"
        '403':
          description: Can only check own collection status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No collection status found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /auth/login (24h validity).
        Used by both manual API calls and the automatic scheduling service.
  parameters:
    UsernameParam:
      name: username
      in: path
      required: true
      description: GitHub username
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]([a-zA-Z0-9\-]){0,38}$'
        example: "octocat"
  schemas:
...
    RepositoryCollectionStartedResponse:
      type: object
      properties:
        message:
          type: string
          example: "Repository collection started"
        status:
          type: string
          enum: ["in_progress"]
        summary:
          type: object
          properties:
            username:
              type: string
            message:
              type: string
              example: "Collection in progress. Use GET /repositories/{username}/status to check progress."
        metadata:
          type: object
          properties:
            startedAt:
              type: string
              format: date-time
            estimatedCompletion:
              type: string
              format: date-time
            currentStep:
              type: integer
              minimum: 1
            totalSteps:
              type: integer
              minimum: 1
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum:
            - "VALIDATION_ERROR"
            - "AUTHENTICATION_ERROR"
            - "AUTHORIZATION_ERROR"
            - "NOT_FOUND_ERROR"
            - "EXTERNAL_SERVICE_ERROR"
            - "RATE_LIMIT_ERROR"
            - "INTERNAL_SERVER_ERROR"
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "The provided data does not match the expected format"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "githubToken"
              message:
                type: string
                example: "Username is required"
              code:
                type: string
                example: "too_small"
        timestamp:
          type: string
          format: date-time
          example: "2025-07-27T10:18:08.760Z"
        path:
          type: string
          example: "/api/auth/login"
        method:
          type: string
          example: "POST"
        stack:
          type: string
          description: Error stack trace (development only)
...